Optimize size of mingaplen from uint32_t to unit16_t

+++ pilight-usb-nano.ino
@@ -45,11 +45,11 @@
 #ifdef DEFAULT_RX_SETTINGS
 uint32_t minrawlen =    20;
 uint32_t maxrawlen =   200;
-uint32_t mingaplen =   300;         // Used and showing multiplied by 10
+uint16_t mingaplen =   300;         // Minimum length of footer pulse and maximum length of previous pulses. Used and showing multiplied by 10
 #else
 uint32_t minrawlen =  1000;
 uint32_t maxrawlen =     0;
-uint32_t mingaplen = 10000;         // Used and showing multiplied by 10
+uint16_t mingaplen = 10000;         // Used and showing multiplied by 10
 #endif 
 
 uint16_t maxgaplen =  8200;         // v2 used and change from 5100 to 8200 (for quigg_gt7000 protocol) set and show multiplied by 10
@@ -91,7 +91,7 @@ void setup() {
 
 #ifdef BOOT_SHOW_SETTINGS
   // Show settings at boot, like as: 'v:20,200,3000,82000,2,1,1600@'
-  sprintf(data, "v:%lu,%lu,%lu,%lu,%d,%d,%d@", minrawlen, maxrawlen, mingaplen*10, uint32_t(maxgaplen)*10, VERSION, MIN_PULSELENGTH/10, MAX_PULSELENGTH/10);
+  sprintf(data, "v:%lu,%lu,%lu,%lu,%d,%d,%d@", minrawlen, maxrawlen, uint32_t(mingaplen)*10, uint32_t(maxgaplen)*10, VERSION, MIN_PULSELENGTH/10, MAX_PULSELENGTH/10);
 #ifdef ADD_LINE_FEED
   Serial.println(data);
 #else
@@ -144,7 +144,7 @@ void receive() {
                     maxrawlen = atol(&data[s]);
                 }
                 if(x == 2) {
-                    mingaplen = atol(&data[s])/10;
+                    mingaplen = uint16_t(atol(&data[s])/10);
                 }
                 x++;
                 s = i+1;
@@ -156,7 +156,7 @@ void receive() {
         /*
          * Once we tuned our firmware send back our settings + fw version
          */
-        sprintf(data, "v:%lu,%lu,%lu,%lu,%d,%d,%d@", minrawlen, maxrawlen, mingaplen*10, uint32_t(maxgaplen)*10, VERSION, MIN_PULSELENGTH/10, MAX_PULSELENGTH/10);
+        sprintf(data, "v:%lu,%lu,%lu,%lu,%d,%d,%d@", minrawlen, maxrawlen, uint32_t(mingaplen)*10, uint32_t(maxgaplen)*10, VERSION, MIN_PULSELENGTH/10, MAX_PULSELENGTH/10);
 #ifdef ADD_LINE_FEED
         Serial.println(data);
 #else
